requirements.txt
flask
opencv-python
numpy
face-recognition
pillow
gunicorn
